pyproject.toml
[project]
authors = [{name = "AmadFat", email = "1943269060@qq.com"}]
name = "nnaf-data"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "nvidia-dali-cuda120>=1.49.0",
    "webdataset>=0.2.111",
    "xxhash>=3.5.0",
    "joblib>=1.5.0",
    "nnaf-utils==0.1.0",
    "alive-progress>=3.2.0",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.uv.sources]
nnaf-utils = { git = "https://kkgithub.com/AmadFat/nnaf-utils" }

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "I",
    "SIM",
    "D"
]
ignore = [
    "D100",
    "D102",
    "F401",
    "F403",
    "F405",
]
